requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorlog==6.9.0
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
dm-tree==0.1.9
einops==0.8.1
executing==2.2.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
gast==0.6.0
hf-xet==1.1.8
huggingface-hub==0.34.4
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
lightning-utilities==0.15.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.4
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-cusparselt-cu12==0.7.1
nvidia-dali-cuda120==1.51.2
nvidia-nccl-cu12==2.20.5
nvidia-nvcomp-cu12==5.0.0.6
nvidia-nvimgcodec-cu12==0.6.0.32
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvjpeg-cu12==12.4.0.76
nvidia-nvjpeg2k-cu12==0.9.0.43
nvidia-nvtiff-cu12==0.5.1.75
nvidia-nvtx-cu12==12.4.99
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
portalocker==3.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyg-lib==0.4.0+pt24cu124
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
terminaltables==3.1.10
threadpoolctl==3.6.0
timm==1.0.19
tokenizers==0.21.4
torch==2.4.1+cu124
torch-geometric==2.6.1
torch-kmeans==0.2.0
torch_cluster==1.6.3+pt24cu124
torch_scatter==2.1.2+pt24cu124
torch_sparse==0.6.18+pt24cu124
torch_spline_conv==1.2.2+pt24cu124
torchaudio==2.4.1+cu124
torchmetrics==1.8.1
torchvision==0.19.1+cu124
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
triton==3.0.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
wrapt==1.17.3
yarl==1.20.1
